Add label to Input, factor out InputField
