--- es/components/Form/InputFancy.js
@@ -1,302 +1,0 @@
-var _typeof = typeof Symbol === "function" && typeof Symbol.iterator === "symbol" ? function (obj) { return typeof obj; } : function (obj) { return obj && typeof Symbol === "function" && obj.constructor === Symbol && obj !== Symbol.prototype ? "symbol" : typeof obj; };
-
-var _templateObject = _taggedTemplateLiteralLoose(['\n  position: relative;\n  box-sizing: border-box;\n\n  width: ', ';\n\n  margin: 0;\n\n  background: transparent;\n\n  @media (max-width: 831px) {\n    height: 48px;\n  }\n  @media (min-width: 832px) {\n    height: 56px;\n  }\n'], ['\n  position: relative;\n  box-sizing: border-box;\n\n  width: ', ';\n\n  margin: 0;\n\n  background: transparent;\n\n  @media (max-width: 831px) {\n    height: 48px;\n  }\n  @media (min-width: 832px) {\n    height: 56px;\n  }\n']),
-    _templateObject2 = _taggedTemplateLiteralLoose(['\n  box-sizing: border-box;\n  display: block;\n\n  position: absolute;\n  width: ', ';\n  left: 0;\n  bottom: 0;\n\n  padding: 10px 16px;\n  padding-left: ', ';\n\n  font-family: ', ';\n  font-weight: ', ';\n  color: ', ';\n\n  background: ', ';\n\n  border: 2px solid transparent;\n  border-color: ', ';\n  border-radius: 5px;\n\n  -webkit-appearance: none;\n  -moz-appearance: none;\n  appearance: none;\n  outline: none;\n\n  -webkit-animation: 1ms void-animation-out;\n\n  &::placeholder {\n    opacity: 1;\n    color: ', ';\n    transition: color 0.3s cubic-bezier(0.165, 0.84, 0.44, 1);\n  }\n\n  @media (max-width: 831px) {\n    font-size: ', ';\n    line-height: ', ';\n  }\n  @media (min-width: 832px) {\n    font-size: ', ';\n    line-height: ', ';\n  }\n'], ['\n  box-sizing: border-box;\n  display: block;\n\n  position: absolute;\n  width: ', ';\n  left: 0;\n  bottom: 0;\n\n  padding: 10px 16px;\n  padding-left: ', ';\n\n  font-family: ', ';\n  font-weight: ', ';\n  color: ', ';\n\n  background: ', ';\n\n  border: 2px solid transparent;\n  border-color: ', ';\n  border-radius: 5px;\n\n  -webkit-appearance: none;\n  -moz-appearance: none;\n  appearance: none;\n  outline: none;\n\n  -webkit-animation: 1ms void-animation-out;\n\n  &::placeholder {\n    opacity: 1;\n    color: ', ';\n    transition: color 0.3s cubic-bezier(0.165, 0.84, 0.44, 1);\n  }\n\n  @media (max-width: 831px) {\n    font-size: ', ';\n    line-height: ', ';\n  }\n  @media (min-width: 832px) {\n    font-size: ', ';\n    line-height: ', ';\n  }\n']),
-    _templateObject3 = _taggedTemplateLiteralLoose(['\n  box-sizing: border-box;\n\n  position: absolute;\n  width: auto;\n  left: ', ';\n  bottom: 10px;\n\n  padding: 4px 6px;\n\n  font-family: ', ';\n  font-weight: ', ';\n  color: ', ';\n\n  overflow: hidden;\n  text-overflow: ellipsis;\n  white-space: nowrap;\n  cursor: ', ';\n\n  transition-property: color, transform;\n  transition-duration: 0.3s;\n  transition-timing-function: cubic-bezier(0.165, 0.84, 0.44, 1);\n\n  @media (max-width: 831px) {\n    font-size: ', ';\n    line-height: ', ';\n\n    background: ', ';\n\n    transform-origin: 0 50%;\n    transform: ', ';\n  }\n\n  @media (min-width: 832px) {\n    font-size: ', ';\n    line-height: ', ';\n\n    background: ', ';\n\n    transform-origin: 0 50%;\n    transform: ', ';\n  }\n'], ['\n  box-sizing: border-box;\n\n  position: absolute;\n  width: auto;\n  left: ', ';\n  bottom: 10px;\n\n  padding: 4px 6px;\n\n  font-family: ', ';\n  font-weight: ', ';\n  color: ', ';\n\n  overflow: hidden;\n  text-overflow: ellipsis;\n  white-space: nowrap;\n  cursor: ', ';\n\n  transition-property: color, transform;\n  transition-duration: 0.3s;\n  transition-timing-function: cubic-bezier(0.165, 0.84, 0.44, 1);\n\n  @media (max-width: 831px) {\n    font-size: ', ';\n    line-height: ', ';\n\n    background: ', ';\n\n    transform-origin: 0 50%;\n    transform: ', ';\n  }\n\n  @media (min-width: 832px) {\n    font-size: ', ';\n    line-height: ', ';\n\n    background: ', ';\n\n    transform-origin: 0 50%;\n    transform: ', ';\n  }\n']),
-    _templateObject4 = _taggedTemplateLiteralLoose(['\n  box-sizing: border-box;\n  display: block;\n\n  position: absolute;\n  left: 0;\n  bottom: 0;\n  width: auto;\n\n  margin: 0;\n\n  font-size: 18px;\n  line-height: 18px;\n\n  color: #bbb;\n\n  background: transparent;\n\n  -webkit-appearance: none;\n  -moz-appearance: none;\n  appearance: none;\n  outline: none;\n\n  @media (max-width: 831px) {\n    padding: 14px 15px 11px 15px;\n  }\n  @media (min-width: 832px) {\n    padding: 15px 16px 14px 16px;\n  }\n'], ['\n  box-sizing: border-box;\n  display: block;\n\n  position: absolute;\n  left: 0;\n  bottom: 0;\n  width: auto;\n\n  margin: 0;\n\n  font-size: 18px;\n  line-height: 18px;\n\n  color: #bbb;\n\n  background: transparent;\n\n  -webkit-appearance: none;\n  -moz-appearance: none;\n  appearance: none;\n  outline: none;\n\n  @media (max-width: 831px) {\n    padding: 14px 15px 11px 15px;\n  }\n  @media (min-width: 832px) {\n    padding: 15px 16px 14px 16px;\n  }\n']);
-
-function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
-
-function _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError("this hasn't been initialised - super() hasn't been called"); } return call && (typeof call === "object" || typeof call === "function") ? call : self; }
-
-function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
-
-function _taggedTemplateLiteralLoose(strings, raw) { strings.raw = raw; return strings; }
-
-import React, { Component, Fragment } from 'react';
-import PropTypes from 'prop-types';
-import styled from 'styled-components';
-
-import { Box } from '../Box';
-
-/**
- * https://medium.com/@ilonacodes/creating-floating-label-placeholder-for-input-with-react-b912233b7005
- *
- * Possible states:
- *  - 1. Empty input that is not focused
- *       Label is transparent (opacity: 0). Use the actual input placeholder.
- *       Label sits on on the input but when focused it will move (transform).
- *  - 2. Empty input that is focused - moves label to top
- *  - 3. Non-empty input - moves label to top
- *
- * To do:
- *  - [] Allow to be `disabled`
- *  - [] Allow to be in `error` state
- *  - [x] Label can be `disabled` to be set to `false` or `_`
- *  - [] Allow to show an `icon` (passed as a `child`). Uses a generic `renderIcon`
- */
-
-var Field = styled(Box)(_templateObject, function (props) {
-  switch (props.renderWidth) {
-    case 'third':
-      return '30%';
-    case 'full':
-    default:
-      return '100%';
-  }
-});
-
-var Input = styled.input(_templateObject2, function (props) {
-  return props.width || '100%';
-}, function (props) {
-  switch (true) {
-    case props.renderIcon:
-      return '48px';
-    default:
-      return '10px';
-  }
-}, function (_ref) {
-  var theme = _ref.theme;
-  return theme.fonts.secondary;
-}, function (_ref2) {
-  var theme = _ref2.theme;
-  return theme.defaults.fontWeight;
-}, function (_ref3) {
-  var theme = _ref3.theme;
-  return theme.colors.grey;
-}, function (_ref4) {
-  var theme = _ref4.theme,
-      isEmpty = _ref4.isEmpty,
-      isFocused = _ref4.isFocused;
-  return isFocused || !isEmpty ? 'white' : theme.backgrounds.inputs;
-}, function (_ref5) {
-  var state = _ref5.state,
-      theme = _ref5.theme,
-      renderType = _ref5.renderType,
-      isFocused = _ref5.isFocused;
-
-  switch (true) {
-    case state === 'danger':
-      return theme.colors.danger;
-    case isFocused && renderType === 'host':
-      return theme.colors.host_primary;
-    case isFocused && renderType === 'primary':
-      return _typeof(theme.colors.primary) === 'object' ? theme.colors.shadower.base : theme.colors.primary;
-    default:
-      return theme.colors.grey_lighter;
-  }
-}, function (_ref6) {
-  var isEmpty = _ref6.isEmpty,
-      isFocused = _ref6.isFocused;
-  return isFocused || !isEmpty ? '#cfd7df' : 'transparent';
-}, function (_ref7) {
-  var theme = _ref7.theme;
-  return theme.defaults.fontSizeMobile;
-}, function (_ref8) {
-  var theme = _ref8.theme;
-  return theme.defaults.fontSizeMobile;
-}, function (_ref9) {
-  var theme = _ref9.theme;
-  return theme.defaults.fontSize;
-}, function (_ref10) {
-  var theme = _ref10.theme;
-  return theme.defaults.fontSize;
-});
-
-var Label = styled.label(_templateObject3, function (_ref11) {
-  var renderIcon = _ref11.renderIcon,
-      isFocused = _ref11.isFocused,
-      isEmpty = _ref11.isEmpty;
-
-  switch (true) {
-    case renderIcon && !(isFocused || !isEmpty):
-      return '48px';
-    default:
-      return '8px';
-  }
-}, function (_ref12) {
-  var theme = _ref12.theme;
-  return theme.fonts.secondary;
-}, function (_ref13) {
-  var theme = _ref13.theme;
-  return theme.defaults.fontWeight;
-}, function (_ref14) {
-  var theme = _ref14.theme,
-      isFocused = _ref14.isFocused,
-      renderType = _ref14.renderType;
-
-  switch (true) {
-    case isFocused && renderType === 'host':
-      return theme.colors.host_primary;
-    case isFocused && renderType === 'primary':
-      return _typeof(theme.colors.primary) === 'object' ? theme.colors.shadower.base : theme.colors.primary;
-    default:
-      return theme.colors.grey_light;
-  }
-}, function (_ref15) {
-  var isEmpty = _ref15.isEmpty,
-      isFocused = _ref15.isFocused;
-  return isFocused || !isEmpty ? 'default' : 'text';
-}, function (_ref16) {
-  var theme = _ref16.theme;
-  return theme.defaults.fontSizeMobile;
-}, function (_ref17) {
-  var theme = _ref17.theme;
-  return theme.defaults.fontSizeMobile;
-}, function (_ref18) {
-  var theme = _ref18.theme,
-      isEmpty = _ref18.isEmpty,
-      isFocused = _ref18.isFocused;
-  return isFocused || !isEmpty ? 'linear-gradient(\n           to top,\n           ' + theme.backgrounds.white + ' 0%,\n           ' + theme.backgrounds.white + ' 57%,\n           transparent 58%,\n           transparent 100%\n        );' : 'transparent';
-}, function (_ref19) {
-  var isEmpty = _ref19.isEmpty,
-      isFocused = _ref19.isFocused;
-  return isFocused || !isEmpty ? 'scale(0.70) translateY(-32px)' : 'none';
-}, function (_ref20) {
-  var theme = _ref20.theme;
-  return theme.defaults.fontSize;
-}, function (_ref21) {
-  var theme = _ref21.theme;
-  return theme.defaults.fontSize;
-}, function (_ref22) {
-  var theme = _ref22.theme,
-      isEmpty = _ref22.isEmpty,
-      isFocused = _ref22.isFocused;
-  return isFocused || !isEmpty ? 'linear-gradient(\n           to top,\n           ' + theme.backgrounds.white + ' 0%,\n           ' + theme.backgrounds.white + ' 53%,\n           transparent 54%,\n           transparent 100%\n        );' : 'transparent';
-}, function (_ref23) {
-  var isEmpty = _ref23.isEmpty,
-      isFocused = _ref23.isFocused;
-  return isFocused || !isEmpty ? 'scale(0.65) translateY(-38px)' : 'none';
-});
-
-var IconSpan = styled.span(_templateObject4);
-
-var InputFancy = function (_Component) {
-  _inherits(InputFancy, _Component);
-
-  function InputFancy(props) {
-    _classCallCheck(this, InputFancy);
-
-    var _this = _possibleConstructorReturn(this, _Component.call(this, props));
-
-    _this.onChange = function (event) {
-      event.preventDefault();
-      event.persist();
-      _this.props.handleOnChange(event);
-      _this.setState({
-        empty: event.target.value === ''
-      });
-    };
-
-    _this.onFocus = function (_) {
-      _this.setState({
-        focused: !_this.state.focused
-      });
-    };
-
-    _this.onBlur = function (event) {
-      _this.setState({
-        focused: false,
-        empty: event.target.value === ''
-      });
-    };
-
-    _this.state = {
-      focused: false,
-      empty: props.value === ''
-    };
-    return _this;
-  }
-
-  InputFancy.prototype.render = function render() {
-    var _props = this.props,
-        type = _props.type,
-        name = _props.name,
-        label = _props.label,
-        value = _props.value,
-        renderWidth = _props.renderWidth,
-        renderType = _props.renderType,
-        renderIcon = _props.renderIcon,
-        children = _props.children,
-        inlineChildren = _props.inlineChildren;
-    var _state = this.state,
-        focused = _state.focused,
-        empty = _state.empty;
-
-
-    return React.createElement(
-      Fragment,
-      null,
-      React.createElement(
-        Field,
-        { renderWidth: renderWidth },
-        React.createElement(Input, {
-          type: type,
-          name: name,
-          'data-cy': 'input-' + name,
-          placeholder: '',
-          onFocus: this.onFocus,
-          isFocused: focused,
-          onBlur: this.onBlur,
-          onChange: this.onChange,
-          value: value,
-          isEmpty: empty,
-          renderType: renderType,
-          renderIcon: renderIcon
-        }),
-        label && label !== '_' && React.createElement(
-          Label,
-          {
-            htmlFor: name,
-            isFocused: focused,
-            isEmpty: empty,
-            renderType: renderType,
-            renderIcon: renderIcon,
-            onClick: this.onFocus
-          },
-          label
-        ),
-        renderIcon && React.createElement(
-          IconSpan,
-          null,
-          children
-        )
-      ),
-      inlineChildren
-    );
-  };
-
-  return InputFancy;
-}(Component);
-
-InputFancy.propTypes = process.env.NODE_ENV !== "production" ? {
-  // Input mechanics
-  type: PropTypes.string,
-  name: PropTypes.string.isRequired, // also used for id
-  value: PropTypes.string.isRequired,
-  label: PropTypes.string.isRequired,
-  // State of field
-  // state: PropTypes.string,
-  // disabled: PropTypes.bool,
-  // Handlers
-  handleOnChange: PropTypes.func.isRequired,
-  // Rendering
-  renderWidth: PropTypes.string,
-  renderType: PropTypes.string,
-  renderIcon: PropTypes.bool,
-  inlineChildren: PropTypes.node
-} : {};
-
-InputFancy.defaultProps = {
-  // disabled: false,
-  type: 'text',
-  // state: 'normal',
-  renderWidth: 'full',
-  renderType: 'primary',
-  renderIcon: false,
-  inlineChildren: null
-};
-
-InputFancy.displayName = 'InputFancy';
-
-export default InputFancy;
