+++ es/components/Form/Label.js
@@ -1,6 +1,6 @@
 var _extends = Object.assign || function (target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i]; for (var key in source) { if (Object.prototype.hasOwnProperty.call(source, key)) { target[key] = source[key]; } } } return target; };
 
-var _templateObject = _taggedTemplateLiteralLoose(['\n  display: inline-block;\n  vertical-align: middle;\n\n  ', '\n  ', '\n  ', '\n'], ['\n  display: inline-block;\n  vertical-align: middle;\n\n  ', '\n  ', '\n  ', '\n']);
+var _templateObject = _taggedTemplateLiteralLoose(['\n  display: inline-block;\n  vertical-align: middle;\n\n  ', '\n  ', '\n  ', '\n  width: 100%;\n  flex: 1 1 100%;\n'], ['\n  display: inline-block;\n  vertical-align: middle;\n\n  ', '\n  ', '\n  ', '\n  width: 100%;\n  flex: 1 1 100%;\n']);
 
 function _taggedTemplateLiteralLoose(strings, raw) { strings.raw = raw; return strings; }
